@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --env=dev --content-base --inline",
+    "start": "webpack serve --config webpack.dev.js",
     "build": "webpack --env=prod -p",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -25,33 +25,39 @@
     "@babel/preset-react": "^7.10.1",
     "babel-loader": "^8.1.0",
     "core-js": "^3.6.5",
-    "css-loader": "^3.5.3",
+    "css-loader": "^6.6.0",
     "file-loader": "^6.0.0",
-    "html-loader": "^1.1.0",
-    "html-webpack-plugin": "^4.3.0",
-    "image-webpack-loader": "^6.0.0",
-    "mini-css-extract-plugin": "^0.9.0",
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "image-webpack-loader": "^8.1.0",
+    "mini-css-extract-plugin": "^2.6.0",
     "regenerator-runtime": "^0.13.7",
-    "stylus": "^0.54.7",
-    "stylus-loader": "^3.0.2",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0"
+    "stylus": "^0.56.0",
+    "stylus-loader": "^6.2.0",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
+    "@mui/icons-material": "^5.4.4",
+    "@mui/material": "^5.4.4",
+    "@mui/styles": "^5.4.4",
     "@reach/router": "^1.3.3",
     "@sentry/apm": "^5.21.3",
-    "@sentry/react": "^5.21.3",
+    "@sentry/react": "^6.18.1",
     "classnames": "^2.2.6",
     "colorthief": "^2.3.0",
-    "i18next": "^19.5.4",
+    "i18next": "^21.6.13",
     "lodash-es": "^4.17.15",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-cropper": "^1.3.0",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-cropper": "^2.1.8",
+    "react-dom": "^17.0.2",
     "react-i18next": "^11.7.0",
     "react-redux": "^7.2.0",
     "redux": "^4.0.5",
