@@ -3,8 +3,8 @@ import { useSelector, useDispatch } from 'react-redux'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 
-import Button from '@material-ui/core/Button'
-import Modal from '@material-ui/core/Modal'
+import Button from '@mui/material/Button'
+import Modal from '@mui/material/Modal'
 
 import { updateImageState } from '../../actions/image'
 
